backend/pyproject.toml
[project]
name = "rag-admin"
version = "0.1.0"
description = "Web application for creating and managing RAG pipelines"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "authlib>=1.3.0",
    "httpx>=0.26.0",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.6",
    "email-validator>=2.3.0",
    "bcrypt>=5.0.0",
    "itsdangerous>=2.2.0",
    # -------------------------------------------------------------------------
    # OpenTelemetry - Observability instrumentation
    # -------------------------------------------------------------------------
    # Core OTel packages
    "opentelemetry-api>=1.20.0", # API interfaces (what you code against)
    "opentelemetry-sdk>=1.20.0", # SDK implementation (does the actual work)
    # Auto-instrumentation packages (zero-code tracing)
    "opentelemetry-instrumentation-fastapi>=0.41b0", # Traces HTTP requests
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0", # Traces DB queries
    "opentelemetry-instrumentation-httpx>=0.41b0", # Traces outgoing HTTP
    # Exporter to send data to OTel Collector
    "opentelemetry-exporter-otlp>=1.20.0",
    # Structured JSON logging with trace correlation
    "python-json-logger>=2.0.7",
    # -------------------------------------------------------------------------
    # Document Processing
    # -------------------------------------------------------------------------
    "llama-index>=0.10.0", # Core LlamaIndex for document parsing
    "llama-index-readers-file>=0.1.0", # File readers for various formats
    "aiofiles>=23.2.0", # Async file operations
    "python-magic>=0.4.27", # MIME type detection from content
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "aiosqlite>=0.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
